requirements.txt
amqp==2.5.0
billiard==3.6.0.0
celery==4.3.0
Jinja2==2.10.1
kombu==4.6.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.17
mysqlclient==1.4.2.post1
protobuf==3.9.0
PyMySQL==0.9.3
pytz==2019.1
redis==3.2.1
repoze.lru==0.7
Routes==2.4.1
six==1.12.0
SQLAlchemy==1.3.6
SQLAlchemy-Utils==0.34.1
vine==1.3.0
youtube-dl==2019.6.21
ffmpeg
django-environ
